update test structure and flow
